refactor: split transaction and foundry tiles
