--- packages/shared/components/atoms/tiles/ActivityTile.svelte
@@ -1,175 +1,0 @@
-<script lang="typescript">
-    import { time } from '@core/app'
-    import { localize } from '@core/i18n'
-    import {
-        Activity,
-        ActivityAsyncStatus,
-        ActivityDirection,
-        claimActivity,
-        ActivityType,
-        rejectActivity,
-        InclusionState,
-        VerificationStatus,
-        selectedAccountAssets,
-        getAssetFromPersistedAssets,
-        IPersistedAsset,
-    } from '@core/wallet'
-    import { truncateString } from '@lib/helpers'
-    import { closePopup, openPopup } from '@lib/popup'
-    import { ActivityAsyncStatusPill, ClickableTile, HR, Icon, Text, Spinner, AssetIcon } from 'shared/components'
-    import { FontWeight } from 'shared/components/Text.svelte'
-
-    export let activity: Activity
-
-    let asset: IPersistedAsset
-
-    $: activity?.data.assetId, $selectedAccountAssets, (asset = getAssetFromPersistedAssets(activity?.data.assetId))
-    $: title = activity?.getTitle()
-    $: subject = activity?.getFormattedSubject()
-    $: isIncomingActivityUnclaimed =
-        activity?.data.type === ActivityType.Transaction &&
-        (activity?.data.direction === ActivityDirection.In || activity?.data.isSelfTransaction) &&
-        activity?.data.asyncStatus === ActivityAsyncStatus.Unclaimed
-    $: timeDiff = activity?.getTimeDiffUntilExpirationTime($time)
-    $: description = getDescription(activity)
-
-    function handleTransactionClick(): void {
-        if (asset?.verification === VerificationStatus.New) {
-            openPopup({
-                type: 'tokenInformation',
-                overflow: true,
-                props: {
-                    activity,
-                    asset,
-                },
-            })
-        } else {
-            openPopup({
-                type: 'activityDetails',
-                props: { activityId: activity.id },
-            })
-        }
-    }
-
-    function handleRejectClick(): void {
-        openPopup({
-            type: 'confirmation',
-            props: {
-                title: localize('actions.confirmRejection.title'),
-                description: localize('actions.confirmRejection.description'),
-                hint: localize('actions.confirmRejection.node'),
-                warning: true,
-                confirmText: localize('actions.reject'),
-                onConfirm: () => {
-                    rejectActivity(activity?.id)
-                    closePopup()
-                },
-            },
-        })
-    }
-
-    function handleClaimClick(): void {
-        claimActivity(activity)
-    }
-
-    function getDescription(activity: Activity): string {
-        if (activity?.type === ActivityType.Foundry) {
-            return asset?.metadata?.name
-                ? truncateString(asset?.metadata?.name, 20, 0)
-                : truncateString(asset?.id, 6, 7)
-        } else if (activity?.data.type === ActivityType.Transaction) {
-            return localize(
-                activity?.data.direction === ActivityDirection.In ? 'general.fromAddress' : 'general.toAddress',
-                { values: { account: subject } }
-            )
-        }
-    }
-</script>
-
-{#if activity}
-    <ClickableTile
-        onClick={handleTransactionClick}
-        classes={activity?.inclusionState !== InclusionState.Confirmed ? 'opacity-50' : ''}
-    >
-        <div class="w-full flex flex-col space-y-4">
-            <div class="flex flex-row items-center text-left space-x-4">
-                <AssetIcon {asset} showVerifiedBadgeOnly />
-                <div class="flex flex-col w-full space-y-0.5">
-                    <div class="flex flex-row justify-between space-x-1">
-                        <Text
-                            fontWeight={FontWeight.semibold}
-                            lineHeight="140"
-                            classes="overflow-hidden overflow-ellipsis multiwrap-line2"
-                        >
-                            {localize(title)}
-                        </Text>
-                        <Text
-                            fontWeight={FontWeight.semibold}
-                            lineHeight="140"
-                            color={activity?.data.type === ActivityType.Transaction &&
-                            activity?.data.direction === ActivityDirection.In
-                                ? 'blue-700'
-                                : ''}
-                            classes="whitespace-nowrap"
-                        >
-                            {activity?.getFormattedAmount(true)}
-                        </Text>
-                    </div>
-
-                    <div class="flex flex-row justify-between">
-                        <Text fontWeight={FontWeight.normal} lineHeight="140" color="gray-600">
-                            {description}
-                        </Text>
-                        <Text
-                            fontWeight={FontWeight.normal}
-                            lineHeight="140"
-                            color="gray-600"
-                            classes="whitespace-nowrap"
-                        >
-                            {activity?.getFiatAmount()}
-                        </Text>
-                    </div>
-                </div>
-            </div>
-            {#if activity.data.type === ActivityType.Transaction && activity?.data.isAsync && (activity.data.direction === ActivityDirection.Out || !activity?.data.isClaimed)}
-                <HR />
-                <div class="flex w-full justify-between space-x-4">
-                    <div class="flex flex-row justify-center items-center space-x-2">
-                        {#if !activity?.data.isClaimed}
-                            <Icon width="16" height="16" icon="timer" classes="text-gray-600" />
-                            <Text fontSize="13" color="gray-600" fontWeight={FontWeight.semibold}
-                                >{timeDiff ?? localize('general.none')}</Text
-                            >
-                        {/if}
-                    </div>
-                    <div class="flex flex-row justify-end w-1/2 space-x-2">
-                        {#if isIncomingActivityUnclaimed}
-                            <button
-                                disabled={activity.data.isClaiming || activity.data.isRejected}
-                                class="action px-3 py-1 w-1/2 text-center rounded-4 font-normal text-14 text-blue-500 bg-transparent 
-                                {activity.data.isClaiming || activity.data.isRejected
-                                    ? 'cursor-default text-gray-500'
-                                    : 'cursor-pointer hover:bg-blue-200'}"
-                                on:click|stopPropagation={handleRejectClick}
-                            >
-                                {localize('actions.reject')}
-                            </button>
-                            <button
-                                class="action px-3 py-1 w-1/2 h-8 text-center rounded-4 font-normal text-14 text-white bg-blue-500 hover:bg-blue-600 dark:hover:bg-blue-400"
-                                on:click|stopPropagation={handleClaimClick}
-                            >
-                                {#if activity.data.isClaiming}
-                                    <Spinner busy={true} classes="justify-center h-fit" />
-                                {:else}
-                                    {localize('actions.claim')}
-                                {/if}
-                            </button>
-                        {:else}
-                            <ActivityAsyncStatusPill asyncStatus={activity.data.asyncStatus} />
-                        {/if}
-                    </div>
-                </div>
-            {/if}
-        </div>
-    </ClickableTile>
-{/if}
