+++ packages/shared/lib/core/wallet/utils/outputs/index.ts
@@ -8,6 +8,7 @@ export * from './getRecipientAddressFromOutput'
 export * from './getRecipientFromOutput'
 export * from './getStorageDepositFromOutput'
 export * from './getTagFromOutput'
+export * from './getTimeDifferenceUntilExpirationTime'
 export * from './getTransactionActivityData'
 export * from './isOutputAsync'
 export * from './preprocessGroupedOutputs'
